requirements.txt
bitsandbytes==0.43.1
datasets==2.19.1
huggingface-hub==0.23.2
hydra-core==1.3.2
nltk==3.8.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp310-cp310-linux_x86_64.whl#sha256=d8cd837ed43e87f77e6efaa08e8de927ca030a1c9c5d04624432d6fb9a74a5ee
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.40.3
pandas==2.2.2
peft==0.11.1
protobuf==3.20.3
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.14.1
sentencepiece==0.2.0
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.4
transformers==4.44.2
triton==2.3.0
trl==0.8.6
typing_extensions==4.12.0
wandb==0.17.0